+++ tasks/mpi/borisov_sum_of_rows/perf_tests/main.cpp
@@ -45,7 +45,7 @@ TEST(borisov_sum_of_rows, test_pipeline_run) {
   perfAnalyzer->pipeline_run(perfAttr, perfResults);
   if (world.rank() == 0) {
     ppc::core::Perf::print_perf_statistic(perfResults);
-    for (int i = 0; i < rows; ++i) {
+    for (size_t i = 0; i < rows; ++i) {
       ASSERT_EQ(global_row_sums[i], cols);
     }
   }
@@ -89,7 +89,7 @@ TEST(borisov_sum_of_rows, test_task_run) {
 
   if (world.rank() == 0) {
     ppc::core::Perf::print_perf_statistic(perfResults);
-    for (int i = 0; i < rows; i++) {
+    for (size_t i = 0; i < rows; i++) {
       ASSERT_EQ(global_row_sums[i], cols);
     }
   }
